+++ R/extract.R
@@ -138,7 +138,15 @@ extract.simGDINA <- function(object,
 #' Objects which can be extracted from \code{GDINA} objects include:
 #'
 #' \describe{
+#' \item{AIC}{AIC}
 #'   \item{att.prior}{attribute prior weights for calculating marginalized likelihood in the last EM iteration}
+#' \item{BIC}{BIC}
+#' \item{catprob.cov}{covariance matrix of item probability parameter estimates; Need to specify \code{SE.type}}
+#' \item{catprob.parm}{item parameter estimates}
+#' \item{catprob.se}{standard error of item probability parameter estimates; Need to specify \code{SE.type}}
+#' \item{delta.cov}{covariance matrix of delta parameter estimates; Need to specify \code{SE.type}}
+#' \item{delta.parm}{delta parameter estimates}
+#' \item{delta.se}{standard error of delta parameter estimates; Need to specify \code{SE.type}}
 #'   \item{discrim}{GDINA discrimination index}
 #'   \item{designmatrix}{A list of design matrices for each item/category}
 #' \item{expectedCorrect}{expected # of examinees in each latent group answering item correctly}
@@ -148,6 +156,7 @@ extract.simGDINA <- function(object,
 #' \item{initial.catprob}{initial item category probability parameters}
 #'   \item{prevalence}{prevalence of each attribute}
 #'   \item{posterior.prob}{posterior weights for each latent class}
+#'   \item{reduced.LG}{Reduced latent group for each item}
 #' }
 #' @param object objects from class \code{GDINA},\code{itemfit}, \code{modelcomp}, \code{Qval} or \code{simGDINA}
 #' @param what what to extract