+++ R/sim10MCDINA.R
@@ -1,0 +1,10 @@
+#' Simulated data (10 items, MC-DINA model)
+#'
+#' Simulated data, Q-matrix and item parameters for a 10-item test measuring 3 attributes.
+#'
+#' @format A list with components:
+#' \describe{
+#' \item{\code{simdat}}{simulated responses of 3000 examinees}
+#' \item{\code{simQ}}{artificial Q-matrix}
+#' }
+"sim10MCDINA"